# Conflicts:
#	app/src/main/java/com/company/khomasi/presentation/favorite/FavouritePage.kt
#	app/src/main/java/com/company/khomasi/presentation/favorite/FavouriteUiState.kt
#	app/src/main/java/com/company/khomasi/presentation/favorite/FavouriteViewModel.kt
#	app/src/main/java/com/company/khomasi/presentation/favorite/MockFavViewModel.kt
#	app/src/main/java/com/company/khomasi/presentation/myBookings/MockViewModel.kt
#	app/src/main/java/com/company/khomasi/presentation/myBookings/MyBookingPage.kt
#	app/src/main/java/com/company/khomasi/presentation/myBookings/MyBookingUiState.kt
#	app/src/main/java/com/company/khomasi/presentation/myBookings/MyBookingViewModel.kt
#	app/src/main/java/com/company/khomasi/presentation/navigation/KhomasiNavigator.kt

+++ app/src/main/java/com/company/khomasi/domain/model/LocalUser.kt
@@ -12,8 +12,8 @@ data class LocalUser(
     val latitude: Double? = null,
     val longitude: Double? = null,
     val profilePicture: String? = null,
-    val coins: Int? = null,
-    val rating: Int? = null,
+    val coins: Double? = null,
+    val rating: Double? = null,
     val token: String? = null,
    // val playgroundId: String? = null,
